+++ projects/demo-app/src/app/scroll-test/scroll-test.module.ts
@@ -3,6 +3,8 @@ import { CommonModule } from '@angular/common';
 import { KtdScrollTestComponent } from './scroll-test.component';
 import { RouterModule, Routes } from '@angular/router';
 import { KtdGridModule } from '@katoid/angular-grid-layout';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
 
 const routes: Routes = [
     {path: 'scroll-test', component: KtdScrollTestComponent},
@@ -15,7 +17,9 @@ const routes: Routes = [
     imports: [
         CommonModule,
         RouterModule.forChild(routes),
-        KtdGridModule
+        KtdGridModule,
+        MatFormFieldModule,
+        MatInputModule
     ]
 })
 export class KtdScrollTestModule {}